Trim unused ToA Points Tracker code
