+++ src/main/java/com/raidtracker/toapointstracker/util/RaidMode.java
@@ -1,4 +1,4 @@
-package com.duckblade.osrs.toa.util;
+package com.raidtracker.toapointstracker.util;
 
 import java.awt.Color;
 import lombok.Getter;
